Improve trigger handling and visualization in oscilloscope
Refactors OscData to use explicit time_min and time_max for windowing. Adds trigger marker drawing to PlotWidget. The trigger marker is now visually indicated on the plot, including offscreen indicators. Added ability to scroll horizontally, meaning trigger is no longer always on the right side of the graph. Trigger can be moved to any location in time on the plot (still need to add a way to actually move it in the ui).