Improved Exporting. Updated Installer. Handled undesired Labrador states.
Exporting now available for all plots. Updated Installer to install drivers for bootloader and gobindar, and updates to include extra firmware files, fonts and images. Changed name of Labrador Controller to LabraScope everywhere. Limited PSU to 11V to prevent entering safety mode. Added checks for if in safety mode telling user to reconnect device. Added check for if in uninitialised state, such as when computer is turned on when board is still plugged in, telling user to reconnect board.